Interaction-Model
This code was written using the help of the Alexa section of the Amazon developer portal. I have to admit I do not know all the inner workings of how this works, but to my knowledge this is the "interaction model" for the Alexa skill. The skill follows a call-response method, meaning it includes a list of "intents." The intents are essentially a list of pre-programmed phrases I chose for the skill, so that when Alexa heard theses it would recognize it was time to speak of the "responses." At this point, the responses is basically just a list of miscellaneous tech facts which I hope will be vaguely useful to users. However, my main problem is eliminating the randomness to the skill: Alexa should be able to deliver advice that is relevant to users' specific tech needs, otherwise its information will be markedly lower in relevancy, and therefore in usefulness. Please help in any way you can, this is part of my independent service project for high school!